// src/app/models/backend/job-application.model.ts
export class JobApplication {
    id?: any;
    job_title?: string;
    job_location?: any;
    resume_file?: any;
    resume_path?: any;
    cover_letter?: any;
    first_name?: any;
    last_name?: any;
    email?: any;
    phone?: any;
    street_address?: any;
    city?: any;
    state_or_province?: any;
    zip_code?: any;
    country?: any;
    linkedIn_profile?: any;
    hear_us_from?: any;
    gender?: any;
    race?: any;
    group_position?: string;
    createdAt?: any;
}